RESTful Routes

+++ app/controllers/cart_controller.rb
@@ -1,7 +1,7 @@
 class CartController < ApplicationController
   before_action :exclude_admin
-  
-  def add_item
+
+  def create
     item = Item.find(params[:item_id])
     session[:cart] ||= {}
     if cart.limit_reached?(item.id)
@@ -17,17 +17,23 @@ class CartController < ApplicationController
   def show
   end
 
-  def empty
-    session.delete(:cart)
-    redirect_to '/cart'
+  def destroy
+    if request.env['PATH_INFO'] == "/cart"
+      session.delete(:cart)
+      redirect_to cart_path
+    else
+      remove_item
+    end
   end
 
   def remove_item
     session[:cart].delete(params[:item_id])
-    redirect_to '/cart'
+    item = Item.find(params[:item_id])
+    flash[:notice] = "#{item.name} has been removed from your cart!"
+    redirect_to cart_path
   end
 
-  def update_quantity
+  def update
     if params[:change] == "more"
       cart.add_item(params[:item_id])
     elsif params[:change] == "less"
@@ -35,6 +41,6 @@ class CartController < ApplicationController
       return remove_item if cart.count_of(params[:item_id]) == 0
     end
     session[:cart] = cart.contents
-    redirect_to '/cart'
+    redirect_to cart_path
   end
 end
